Worked on documentation, refactoring and code clean up. Retested prodigy_dataset_creator.py still working.
