+++ scrc/annotation/requirements.txt
@@ -1,4 +1,6 @@
 psycopg2-binary >=2.7,<2.8
 pandas==1.3.5
 spacy==3.1
+ast
+os
